feat: create a reusable ColorBlock component
