+++ src/App.js
@@ -1,23 +1,9 @@
 import React from 'react'
 import { disableBodyScroll, clearAllBodyScrollLocks } from 'body-scroll-lock'
-import generateSpectrum from './lib/generate-spectrum'
-import colors from './lib/colors/seasonal-colors'
-
-const palette = colors.winter
-
-const spectrum = generateSpectrum(palette, 60)
+import ColorBlock from './components/color-block'
 
 class App extends React.Component {
-
   targetElement = null
-
-  constructor(props) {
-    super(props)
-    this.state = {
-      colors: spectrum
-    }
-    this.changeColor = this.changeColor.bind(this)
-  }
 
   componentDidMount() {
     this.targetElement = document.querySelector('#root')
@@ -28,34 +14,9 @@ class App extends React.Component {
     clearAllBodyScrollLocks()
   }
 
-  changeColor(event) {
-    event.preventDefault()
-    event.stopPropagation()
-
-    // Make a copy of the spectrum.
-    const newColors = this.state.colors.slice()
-    // Move the first element of spectrum to the last.
-    newColors.push(newColors.shift())
-
-    this.setState({
-      colors: newColors
-    })
-  }
-
   render() {
     return (
-      <div
-        onClick={this.changeColor}
-        onTouchStart={this.changeColor}
-        onTouchMove={this.changeColor}
-        onMouseMove={this.changeColor}
-        onTouchEnd={this.changeColor}
-        style={{
-        'minWidth': '100vw',
-        'minHeight': '100vh',
-        'backgroundColor': `#${this.state.colors[0]}`
-      }}>
-      </div>
+      <ColorBlock />
     )
   }
 }
